requirements.txt
Django>=1.11.4
djangorestframework>=3.6
django-cors-headers>=2.1.0